Glowstone and Redstone
Added blocks and functionality for Glowstone golem to glow and Redstone golem to power redstone. Fixed Config accordingly. Added new EntityAI to handle this, but addons that want a glowing or powering golem should still inherit from the respective golems instead of GolemBase if they want the block to work.