Minor Fixes and Start Book
Re-arranged the logic behind the golem build event to enable a Golem Guidebook to be created. It *should* support golems from add-ons because it still fires the event. Added a GUI class in preparation for whenever I figure out how to make the GUI, for now ItemInfoBook uses the book gui.
Also updated texture for bone golem. I left the old one in the file just in case. Also renamed a couple .json files to be lowercase.