Golem Info Book is up to snuff
All basic features have been implemented in the Golem Info Book. Still need to add tabs and refine the Table of Contents and do a bit of localization. And I'm sure some of the code can be streamlined for better performance.
For some reason it requires a world restart to update the language. Also the draw-itemstack thing for the golem diagram doesn't do z-layers correctly. Very odd. For now, the blocks are not actually touching in the diagram.