Containers/Specials Part 2
Finished adapting the new "config-defined special" code to current setup. Golems should check the value of their specials, but this hasn't been tested because the specials are not appearing correctly in the config (they are lumped all in one category but not all of the specials have unique names).
Started fixing BlockUtility and it's children, needs more work.